test: isolate functional tests with per-file server instances
Each test file now gets its own fresh server + database via setupFiles,
eliminating shared state issues. Restores the "or" divider test that was
removed due to federation provider leakage across tests.
