+++ tests/functional/vitest.config.ts
@@ -3,6 +3,7 @@ import { defineConfig } from 'vitest/config';
 export default defineConfig({
   test: {
     globalSetup: ['./setup.ts'],
+    setupFiles: ['./setup-per-file.ts'],
     testTimeout: 15000,
     hookTimeout: 60000,
     fileParallelism: false,
